feat: implemented the camera feature
