--- app/src/main/java/com/tolstoy/zurichat/ui/activities/Open1CameraActivity.java
@@ -1,41 +1,0 @@
-package com.tolstoy.zurichat.ui.activities;
-
-import androidx.activity.result.ActivityResult;
-import androidx.activity.result.ActivityResultCallback;
-import androidx.activity.result.ActivityResultLauncher;
-import androidx.activity.result.contract.ActivityResultContract;
-import androidx.activity.result.contract.ActivityResultContracts;
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.widget.ImageView;
-
-import com.tolstoy.zurichat.R;
-
-public class Open1CameraActivity extends AppCompatActivity {
-
-    public static final int CAMERA_ACTION_CODE = 1;
-    ImageView imageView;
-    ImageView imageButton;
-    ActivityResultLauncher<Intent> activityResultLauncher;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_open_camera);
-
-        imageView = findViewById(R.id.displayid_ImgV);
-        imageButton = findViewById(R.id.camera_btn);
-
-        activityResultLauncher = registerForActivityResult(new ActivityResultContracts.StartActivityForResult(), new ActivityResultCallback<ActivityResult>() {
-            @Override
-            public void onActivityResult(ActivityResult result) {
-
-            }
-        }
-
-        });
-
-    }
-}
